refactor(protocol): split typing-analytics-db into statement modules and a class chain
